# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing HydraDNS server configuration and filtering.",
        "title": "HydraDNS Management API",
        "contact": {
            "name": "HydraDNS Support",
            "url": "https://github.com/jroosing/hydradns"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/custom-dns": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all configured custom DNS host and CNAME records",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "custom-dns"
                ],
                "summary": "List all custom DNS records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.CustomDNSRecordsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/custom-dns/cnames": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Adds a new custom DNS CNAME record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "custom-dns"
                ],
                "summary": "Add a CNAME record",
                "parameters": [
                    {
                        "description": "CNAME record to add",
                        "name": "record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.AddCNAMERequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.CustomDNSOperationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "CNAME already exists",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/custom-dns/cnames/{alias}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the target for an existing custom DNS CNAME",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "custom-dns"
                ],
                "summary": "Update a CNAME record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CNAME alias",
                        "name": "alias",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated target",
                        "name": "record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.UpdateCNAMERequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.CustomDNSOperationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes a custom DNS CNAME record",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "custom-dns"
                ],
                "summary": "Delete a CNAME record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CNAME alias",
                        "name": "alias",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.CustomDNSOperationResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/custom-dns/hosts": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Adds a new custom DNS host record (A/AAAA)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "custom-dns"
                ],
                "summary": "Add a host record",
                "parameters": [
                    {
                        "description": "Host record to add",
                        "name": "record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.AddHostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.CustomDNSOperationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Host already exists",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/custom-dns/hosts/{name}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the IP addresses for an existing custom DNS host",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "custom-dns"
                ],
                "summary": "Update a host record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Host name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated IP addresses",
                        "name": "record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.UpdateHostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.CustomDNSOperationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes a custom DNS host record",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "custom-dns"
                ],
                "summary": "Delete a host record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Host name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.CustomDNSOperationResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_jroosing_hydradns_internal_api_models.AddCNAMERequest": {
            "type": "object",
            "required": [
                "alias",
                "target"
            ],
            "properties": {
                "alias": {
                    "type": "string"
                },
                "target": {
                    "type": "string"
                }
            }
        },
        "github_com_jroosing_hydradns_internal_api_models.AddHostRequest": {
            "type": "object",
            "required": [
                "ips",
                "name"
            ],
            "properties": {
                "ips": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_jroosing_hydradns_internal_api_models.CustomDNSCountsResponse": {
            "type": "object",
            "properties": {
                "cnames": {
                    "type": "integer"
                },
                "hosts": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_jroosing_hydradns_internal_api_models.CustomDNSOperationResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_jroosing_hydradns_internal_api_models.CustomDNSRecordsResponse": {
            "type": "object",
            "properties": {
                "cnames": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "count": {
                    "$ref": "#/definitions/github_com_jroosing_hydradns_internal_api_models.CustomDNSCountsResponse"
                },
                "hosts": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "github_com_jroosing_hydradns_internal_api_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_jroosing_hydradns_internal_api_models.UpdateCNAMERequest": {
            "type": "object",
            "required": [
                "target"
            ],
            "properties": {
                "target": {
                    "type": "string"
                }
            }
        },
        "github_com_jroosing_hydradns_internal_api_models.UpdateHostRequest": {
            "type": "object",
            "required": [
                "ips"
            ],
            "properties": {
                "ips": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}